Update to make more idiomatic
Since there is an emphasis on writing concise code, I think it would be a good idea to have the when expression directly be an argument. It will save a variable definition and is more in line with Kotlin style. Code lab documents probably need to be updated as well but could not find if they were open source.